+++ mycroft/tts/google_tts.py
@@ -14,7 +14,7 @@
 #
 from gtts import gTTS
 
-from mycroft.tts import TTS, TTSValidator
+from .tts import TTS, TTSValidator
 
 
 class GoogleTTS(TTS):
